// syncro/apps/web/src/lib/api/generated/model/auditLogListResponse.ts
/**
 * Generated by orval v8.13.0 🍺
 * Do not edit manually.
 * Syncro API
 * Syncro backend REST API.
 * OpenAPI spec version: v1
 */
import type { AuditLogEntryView } from './auditLogEntryView';

export interface AuditLogListResponse {
  items?: AuditLogEntryView[];
  totalElements?: number;
  page?: number;
  size?: number;
  sort?: string;
}